Better defaults
1st time that a key is inserted, it'll be written in a file and remembered every subsequent time that the plugin is launched and is shown in the corresponding box.  If a new key is inserted, the file will be updated and the new key will be shown hereafter.
When the plugin is launched in a saved Qgis project file, the project file path is retrieved and shown in the corresponding box.
Boxes for selecting OD, Origin, and Destination layers now show only point layers (which is the only acceptable layer type) and have a default layer provided that at least one point layer is loaded in the current Qgis instance.