UI update - Instance is split in 3
Instance is split in 3 as Wim suggested: Client, Network, Instance
Once “client” is filled in for the 1st time, for the subsequent times that the plugin is reloaded, Toolbox remembers and fills in the client box (just like with the Key.
Once a “network” is filled in and used, for the subsequent times that the plugin is reloaded, Toolbox remembers and suggests used networks as soon the starting alphabet is typed (in the image below, I typed “a” and got the corresponding suggestions). This list gets updated upon inserting a new network, and next time the Toolbox is reloaded, this new network will be amongst the suggestions. The list of suggestion will be updated only if a NEW network is typed-in.
The instance box is turned to a dropdown box that consists of only 0-9 to be entirely consistent with our instance numbering and it is not editable to avoid potential typing mistakes.